moving utility functions in pylearn.expr
